lab2/requirements.txt
kaggle
pytorch
torchvision
torchaudio
scikit-learn
pandas
numpy
